More linux/bash fixes

+++ test/Config_test.cpp
@@ -29,55 +29,55 @@ class ConfigTest : public ::testing::Test {
 // parse duration
 
 TEST_F(ConfigTest, ParseDuration_Hours) {
-  long long milliseconds;
+  int64_t milliseconds;
   ConfigTestHelper::ParseResult result = config->parse_duration("1h", milliseconds);
   EXPECT_EQ(result, ConfigTestHelper::ParseResult::Success);
   EXPECT_EQ(milliseconds, 3600000);
 }
 
 TEST_F(ConfigTest, ParseDuration_Minutes) {
-  long long milliseconds;
+  int64_t milliseconds;
   ConfigTestHelper::ParseResult result = config->parse_duration("5m30s", milliseconds);
   EXPECT_EQ(result, ConfigTestHelper::ParseResult::Success);
   EXPECT_EQ(milliseconds, 330000);
 }
 
 TEST_F(ConfigTest, ParseDuration_Complex) {
-  long long milliseconds;
+  int64_t milliseconds;
   ConfigTestHelper::ParseResult result = config->parse_duration("1h27m16s", milliseconds);
   EXPECT_EQ(result, ConfigTestHelper::ParseResult::Success);
   EXPECT_EQ(milliseconds, 5236000);
 }
 
 TEST_F(ConfigTest, ParseDuration_Milliseconds) {
-  long long milliseconds;
+  int64_t milliseconds;
   ConfigTestHelper::ParseResult result = config->parse_duration("500u", milliseconds);
   EXPECT_EQ(result, ConfigTestHelper::ParseResult::Success);
   EXPECT_EQ(milliseconds, 500);
 }
 
 TEST_F(ConfigTest, ParseDuration_Fractional) {
-  long long milliseconds;
+  int64_t milliseconds;
   ConfigTestHelper::ParseResult result = config->parse_duration("0.5h", milliseconds);
   EXPECT_EQ(result, ConfigTestHelper::ParseResult::Success);
   EXPECT_EQ(milliseconds, 1800000);
 }
 
 TEST_F(ConfigTest, ParseDuration_Combined) {
-  long long milliseconds;
+  int64_t milliseconds;
   ConfigTestHelper::ParseResult result = config->parse_duration("1h0.5m", milliseconds);
   EXPECT_EQ(result, ConfigTestHelper::ParseResult::Success);
   EXPECT_EQ(milliseconds, 3630000);
 }
 
 TEST_F(ConfigTest, ParseDuration_InvalidFormat) {
-  long long milliseconds;
+  int64_t milliseconds;
   ConfigTestHelper::ParseResult result = config->parse_duration("1h2m3.4s5u", milliseconds);
   EXPECT_EQ(result, ConfigTestHelper::ParseResult::InvalidFormat);
 }
 
 TEST_F(ConfigTest, ParseDuration_InvalidUnit) {
-  long long milliseconds;
+  int64_t milliseconds;
   ConfigTestHelper::ParseResult result = config->parse_duration("1k", milliseconds);
   EXPECT_EQ(result, ConfigTestHelper::ParseResult::InvalidUnit);
 }
